refactor(desktop-ui/test): consume :core-test TestFixtures, drop local duplicate (Phase 2)
- Extract TestTopologies from desktop-ui's TestFixtures.kt into its own
  file (parallel to the :core/jvmTest split; duplication between :core
  and :desktop-ui TestTopologies stays for now — separate concern).
- Delete desktop-ui/src/test/.../testutil/TestFixtures.kt (the existing
  :core-test dep provides the consolidated replacement).
- Migrate remaining getResourceAsStream callers
  (XMLContextFactoryTest, XMLContextFactoryOutputStreamTest,
  EditToSimulationWorkflowTest) to Resources.read(...).byteInputStream()
  or TestFixtures named methods.

:desktop-ui:test 456 passed + 1 skipped. Zero getResourceAsStream calls
remain under desktop-ui/src/test or core/src/jvmTest.
