+++ desktop-ui/src/test/kotlin/cz/vutbr/fit/interlockSim/xml/XMLContextFactoryTest.kt
@@ -19,6 +19,7 @@ import assertk.assertions.isInstanceOf
 import assertk.assertions.isNotNull
 import assertk.assertions.isSameInstanceAs
 import assertk.assertions.isTrue
+import cz.vutbr.fit.interlockSim.util.Resources
 import cz.vutbr.fit.interlockSim.context.ContextCreationException
 import cz.vutbr.fit.interlockSim.context.DefaultEditingContext
 import cz.vutbr.fit.interlockSim.context.DefaultRailWayNetGrid
@@ -1217,7 +1218,7 @@ class XMLContextFactoryTest : KoinTestBase() {
 	// Helper method to load fixture files from resources
 	private fun getFixtureStream(fileName: String): InputStream {
 		val resourcePath = "/cz/vutbr/fit/interlockSim/xml/fixtures/$fileName"
-		val stream = javaClass.getResourceAsStream(resourcePath)
+		val stream = Resources.read(resourcePath.trimStart('/')).byteInputStream()
 		assertThat(stream)
 			.withMessage("Fixture file should exist: $fileName")
 			.isNotNull()
